Complete Tic Tac Toe Project
This commit adds everything from the html to the js used to power this
tic tac toe game.

The styling was kept relatively basic with just a simple greyscale color
scheme.

The Javascript was designed so that all of the important computations
are done inside of an IIFE so that users can't tamper with things like
score.